change test function location and address comments
